requirements.txt
Flask
Flask-Cors
requests
Pillow
pytesseract
google-cloud-translate